Added Auth guards + Unauthorized Component

+++ src/app/services/storage.service.ts
@@ -1,4 +1,5 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
 
 const USER_KEY = 'auth-user';
 
@@ -6,10 +7,16 @@ const USER_KEY = 'auth-user';
   providedIn: 'root'
 })
 export class StorageService {
-  constructor() {}
+  private isBrowser: boolean;
+
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {
+    this.isBrowser = isPlatformBrowser(this.platformId);
+  }
 
   clean(): void {
-    window.sessionStorage.clear();
+    if (this.isBrowser) {
+      window.sessionStorage.clear();
+    }
   }
 
   public getToken(): string | null {
@@ -17,22 +24,28 @@ export class StorageService {
     return user ? user.token : null;
   }
   
-
   public saveUser(user: any): void {
-    window.sessionStorage.removeItem(USER_KEY);
-    window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    if (this.isBrowser) {
+      window.sessionStorage.removeItem(USER_KEY);
+      window.sessionStorage.setItem(USER_KEY, JSON.stringify(user));
+    }
   }
 
   public getUser(): any {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    if (user) {
-      return JSON.parse(user);
+    if (this.isBrowser) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      if (user) {
+        return JSON.parse(user);
+      }
     }
-    return { role: '' }; // Return a default object with a single role property
+    return { role: '' }; 
   }
 
   public isLoggedIn(): boolean {
-    const user = window.sessionStorage.getItem(USER_KEY);
-    return user !== null;
+    if (this.isBrowser) {
+      const user = window.sessionStorage.getItem(USER_KEY);
+      return user !== null;
+    }
+    return false;
   }
 }
